adding navigation from details fragment to rate dialog fragment. adding customized toolbar to navigate from it.

+++ app/src/main/java/com/emad/movies/data/model/RequestRate.kt
@@ -1,3 +1,3 @@
 package com.emad.movies.data.model
 
-data class RequestRate(val value: Double)
+data class RequestRate(val value: Float)
